requirements.txt
aiohttp==3.8.1
aiohttp-jinja2==1.5
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
backcall==0.2.0
charset-normalizer==2.0.8
colorama==0.4.4
decorator==5.1.0
dicttoxml==1.7.4
frozenlist==1.2.0
idna==3.3
ipython==7.30.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
multidict==5.2.0
numpy==1.21.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.23
Pygments==2.10.0
scikit-learn==1.0.1
scipy==1.7.3
sklearn==0.0
threadpoolctl==3.0.0
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
traitlets==5.1.1
typing-extensions==4.0.0
wcwidth==0.2.5
yarl==1.7.2
